+++ nodes/recipe_handler.py
@@ -22,8 +22,9 @@ from threading import RLock
 from openag_brain import params, services
 from openag_brain.srv import StartRecipe, Empty
 from openag_brain.load_env_var_types import VariableInfo
-from openag_brain.utils import gen_doc_id, read_environment_from_ns
-from std_msgs.msg import String, Float64, Bool 
+from openag_brain.recipe_interpreters import interpret_simple_recipe
+from openag_brain.utils import gen_doc_id, read_environment_from_ns, trace
+from std_msgs.msg import String, Float64, Bool
 
 import pdb
 # Create a tuple constant of valid environmental variables
@@ -60,62 +61,8 @@ PUBLISHERS = {
 # A threshold to compare time values in seconds.
 THRESHOLD = 1
 
-# Turn on logic tracing by making the variable below True.  
-# Output ONLY is written to this node's log file:
-# tail -f ~/.ros/log/latest/environments-environment_1-recipe_handler_1-6.log
-TRACE = False
-def trace(msg, *args):
-    if TRACE:
-        msg = '\nTRACE> ' + msg
-        rospy.logdebug(msg, *args)
-
 
 #------------------------------------------------------------------------------
-def interpret_simple_recipe(recipe, start_time, now_time):
-    """
-    Produces a tuple of ``(variable, value)`` pairs by building up
-    a recipe state from walking through the recipe keyframes
-    """
-    _id = recipe["_id"]
-    operations = recipe["operations"]
-    rospy.loginfo(operations)
-    end_time_relative = operations[-1][0]
-    trace("recipe_handler: interpret_simple_recipe end_time_relative=%s", 
-        end_time_relative)
-    end_time = start_time + end_time_relative
-    # If start time is at some point in the future beyond the threshold
-    if start_time - now_time > THRESHOLD:
-        raise ValueError("Recipes cannot be scheduled for the future")
-    # If there are no recipe operations, immediately start and stop
-    # The recipe.
-    if not len(operations):
-        return (
-            (RECIPE_START.name, _id),
-            (RECIPE_END.name, _id),
-        )
-    if now_time >= (end_time + THRESHOLD):
-        return ((RECIPE_END.name, _id),)
-    if abs(now_time - start_time) < THRESHOLD:
-        return ((RECIPE_START.name, _id),)
-
-    now_relative = (now_time - start_time)
-
-    # Create a state object to accrue recipe setpoint values.
-    state = {}
-    # Build up state up until now_time (inclusive).
-    trace("recipe_handler: interpret_simple_recipe now=%s", now_relative)
-    for timestamp, variable, value in operations:
-        if timestamp > now_relative:
-            break
-        state[variable] = value
-        trace("recipe_handler: interpret_simple_recipe: %s %s %s", 
-            timestamp, variable, value)
-    rospy.loginfo(state)
-    return tuple(
-        (variable, value)
-        for variable, value in state.iteritems()
-    )
-
 
 RECIPE_INTERPRETERS = {
     "simple": interpret_simple_recipe
@@ -233,7 +180,7 @@ class RecipeHandler:
 
     def register_services(self):
         """Register services for instance"""
-        rospy.Service(services.START_RECIPE, StartRecipe, 
+        rospy.Service(services.START_RECIPE, StartRecipe,
             self.start_recipe_service)
         rospy.Service(services.STOP_RECIPE, Empty, self.stop_recipe_service)
         rospy.set_param(
@@ -271,17 +218,17 @@ class RecipeHandler:
             end_doc = end_view.rows[0].value
             trace("recipe_handler: end_doc=%s", end_doc)
             if (end_doc["timestamp"] > start_doc["timestamp"]):
-                trace("recipe_handler: RETURNING: end_time=%s > start_time=%s", 
+                trace("recipe_handler: RETURNING: end_time=%s > start_time=%s",
                     end_doc["timestamp"], start_doc["timestamp"])
                 return
         # Run the recipe
-        trace("recipe_handler: restarting recipe=%s at time=%s", 
+        trace("recipe_handler: restarting recipe=%s at time=%s",
             start_doc["value"], start_doc["timestamp"])
         self.start_recipe_service(
             StartRecipe._request_class(start_doc["value"]),
             start_doc["timestamp"]
         )
-        
+
     def save_recipe_dp(self, variable):
         """
         Save the recipe start/end to the env. data pt. DB, so we can restart
@@ -291,12 +238,12 @@ class RecipeHandler:
             "environment": self.environment,
             "variable": variable,
             "is_desired": True,
-            "value": rospy.get_param(params.CURRENT_RECIPE), 
+            "value": rospy.get_param(params.CURRENT_RECIPE),
             "timestamp": rospy.get_time()
         })
         doc_id = gen_doc_id(rospy.get_time())
         self.env_data_db[doc_id] = doc
-    
+
 
 #------------------------------------------------------------------------------
 # Our ROS node main entry point.  Starts up the node and then waits forever.
@@ -368,6 +315,3 @@ if __name__ == '__main__':
 
         rate.sleep()
         # end of while loop in main
-
-
-
