+++ src/openag_brain/utils.py
@@ -41,3 +41,12 @@ def read_environment_from_ns(namespace):
         )
     environment_id = result.group(1)
     return environment_id
+
+# Turn on logic tracing by making the variable below True.
+# Output ONLY is written to this node's log file:
+# tail -f ~/.ros/log/latest/environments-environment_1-recipe_handler_1-6.log
+TRACE = False   # TODO: Move to settings file?
+def trace(msg, *args):
+    if TRACE:
+        msg = '\nTRACE> ' + msg
+        rospy.logdebug(msg, *args)
